refactor(web): consolidate landing image and type modules
