+++ apps/web/components/Feature/landing/Hero/index.tsx
@@ -24,7 +24,7 @@ import { PATHS } from "@/utils/path";
 import ScrollReveal from "@/components/UI/ScrollReveal";
 import ImageReveal from "@/components/UI/ImageReveal";
 import { LANDING_REVEAL } from "@/utils/landingUtils";
-import { IMAGE_SIZES } from "@/utils/imageSizes";
+import { IMAGE_SIZES } from "@/utils/landingShared";
 import {
   LANDING_IMAGE_FLAGS,
   LANDING_REVEAL_VARIANTS,
