+++ apps/web/components/Feature/landing/SecurePayment/index.tsx
@@ -20,7 +20,7 @@ import COLORS from "@repo/ui/colors";
 import ScrollReveal from "@/components/UI/ScrollReveal";
 import ImageReveal from "@/components/UI/ImageReveal";
 import { LANDING_REVEAL } from "@/utils/landingUtils";
-import { IMAGE_SIZES } from "@/utils/imageSizes";
+import { IMAGE_SIZES } from "@/utils/landingShared";
 import {
   LANDING_IMAGE_FLAGS,
   LANDING_REVEAL_VARIANTS,
