+++ apps/web/components/Feature/landing/Testimonial/index.tsx
@@ -19,7 +19,7 @@ import { MonoText } from "@/components/UI/Monotext";
 import ScrollReveal from "@/components/UI/ScrollReveal";
 import ImageReveal from "@/components/UI/ImageReveal";
 import { LANDING_REVEAL } from "@/utils/landingUtils";
-import { IMAGE_SIZES } from "@/utils/imageSizes";
+import { IMAGE_SIZES } from "@/utils/landingShared";
 import {
   LANDING_IMAGE_FLAGS,
   LANDING_REVEAL_VARIANTS,
