+++ apps/web/providers/smoothScrollProvider.tsx
@@ -4,7 +4,7 @@ import Lenis from "lenis";
 import { useEffect } from "react";
 import { gsap } from "gsap";
 import { ScrollTrigger } from "gsap/ScrollTrigger";
-import { type SmoothScrollProviderProps } from "@/types/smoothScrollProvider";
+import { type SmoothScrollProviderProps } from "@/utils/landingShared";
 import { SMOOTH_SCROLL, SMOOTH_SCROLL_EVENTS } from "@/utils/landingUtils";
 
 if (typeof window !== "undefined") {
@@ -128,4 +128,3 @@ export function SmoothScrollProvider({ children }: SmoothScrollProviderProps) {
 
   return children;
 }
-
