--- apps/web/types/discoverCard.ts
@@ -1,6 +1,0 @@
-import type { DiscoverContentItem } from "@/utils/discoverContent";
-
-export type DiscoverCardProps = {
-  item: DiscoverContentItem;
-  lng?: string;
-};
