--- apps/web/types/safeImage.ts
@@ -1,8 +1,0 @@
-import { type ImageProps, type StaticImageData } from "next/image";
-
-export type SafeImageSrc = string | StaticImageData;
-
-export type SafeImageProps = Omit<ImageProps, "src"> & {
-  src: SafeImageSrc;
-  fallback?: SafeImageSrc;
-};
