--- apps/web/types/scrollReveal.ts
@@ -1,6 +1,0 @@
-import { type HTMLAttributes, type ReactNode } from "react";
-
-export interface ScrollRevealProps extends HTMLAttributes<HTMLDivElement> {
-  children: ReactNode;
-  delay?: number;
-}
