--- apps/web/types/smoothScrollProvider.ts
@@ -1,5 +1,0 @@
-import { type ReactNode } from "react";
-
-export interface SmoothScrollProviderProps {
-  children: ReactNode;
-}
