--- apps/web/types/valueStatement.ts
@@ -1,9 +1,0 @@
-import { type ImageSource } from "@/utils/Constants";
-
-export type ValueStatementProps = {
-  bgImage?: ImageSource;
-  title?: string;
-  subtitle?: string;
-  ctaText?: string;
-  ctaHref?: string;
-};
