--- apps/web/utils/imageReveal.ts
@@ -1,11 +1,0 @@
-export const IMAGE_REVEAL_DEFAULTS = {
-  variant: "fade-scale",
-  delay: 0,
-  duration: 1.4,
-  start: "top 88%",
-  clipRevealVariant: "clip-reveal",
-  willChange: "opacity, transform, clip-path",
-  toggleActions: "play none none reverse",
-  kenBurnsScaleFrom: 1.06,
-  scaleDefaultFrom: 0.96,
-} as const;
