--- apps/web/utils/imageSizes.ts
@@ -1,6 +1,0 @@
-export const IMAGE_SIZES = {
-  fullViewport: "100vw",
-  securePayment: "(max-width: 1024px) 100vw, 42vw",
-  discoverCard: "(max-width: 767px) 100vw, (max-width: 1199px) 50vw, 33vw",
-  genericCard: "(max-width: 767px) 100vw, 50vw",
-} as const;
